Update used JDKs to Java 8.
As Graylog 2.0 will require a Java 8 runtime, plugins should (although
this is not required) be written and built on Java 8. Therefore this
commit changes the compiler's target and language level, as well als
travis' build config to Java 8.

This should be merged either to a branch in preparation of 2.0 or when
Graylog2/graylog2-server#1479 is merged.